refactor: Обновил конфигурационный файл, провел работу над менеджментом памяти, убрал не используемые функции.

+++ src/main.h
@@ -9,8 +9,8 @@
 struct ProgramOptions {
     std::filesystem::path directory_path = "./"; 
     std::vector<std::string> ignore_patterns;
-    int max_recursion_depth = 1; // Default value: 1, safe mod
-    int char_style = 2;           // Default value: ╠═, ╚═, ║
+    short int max_recursion_depth = 1; // Default value: 1, safe mod
+    short int char_style = 2;           // Default value: ╠═, ╚═, ║
     bool tree_style = false;      // Default value: directories first
     bool need_help = false;       // Default value: no help needed
     bool need_config = false;     // Default value: no config needed
